// src/utils/types/FormSection.ts
export enum FormSectionOptions {
    BASIC_INFO = "Definições Básicas",
    RESOURCES = "Recursos",
    PARTS = "Partes",
    EVALUATION_CRITERIA = "Critérios de Avaliação",
    EVALUATION_METHOD = "Forma de Avaliação",
    CALCULATION = "Cálculo"
}

export enum SequenceFormOptions {
    BASIC_INFO = 'Definições básicas',
    OBJECTIVES = 'Objetivos',
    SKILLS = 'Habilidades',
    BNCC_CODES ='Códigos BNCC',
    STAGES = 'Fases'
}